StrUtils: fixed address parsing attempting to resolve strings
By default, if InetAddress.getByName() is given a string that is not a
valid address it will attempt to resolve the string. This is problematic
with DNS servers, such as OpenDNS, that will always return an address,
even for invalid hostnames/domains. Moreover, there is absolutely no
point in attempting resolve invalid addresses as the purpose is to
validate the address.

This fixes the issue of some players appearing as addresses when actions
are applied to them.